requirements.txt
colorama==0.4.4
Cython==0.29.24
dominate==2.6.0
future==0.18.2
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==10.0.0
protobuf==3.17.3
PyYAML==5.4.1
scipy==1.7.1
tensorboardX==2.4
torch==2.0.1
torchvision==0.15.2
tqdm==4.62.2
visdom==0.1.8.9
yacs==0.1.8

argparse~=1.4.0
utils~=1.0.1
pandas~=2.0.0
scikit-image~=0.21.0
requests~=2.22.0
beautifulsoup4~=4.12.2
